fix: uppy doesn't works well
