+++ app/views/speaker_dashboard/speakers/_form_speaker.html.erb
@@ -40,12 +40,13 @@
   <%= form.text_field :github_id, class: "form-control", placeholder: 'cloudnativedaysjp'  %>
 </div>
 
-<div class="field pb-3">
+<div class="field pb-3" data-controller="crop-upload">
   <%= form.label :avatar, '講演者写真 - Photographs' %>
   <!-- will be replaced by Uppy -->
   <%= form.text_field :avatar,
                       type: :file,
                       id: "avatar_upload",
+                      "data-crop-upload-target": "fileInput",
                       class: "form-control" %>
 
   <!-- for storing uploaded file data -->
